refactor: move addStone function into child
